Forge/LakeIngestor--Events__DEV@v0.1.0/requirements.txt
confluent-kafka==2.3.0
pyarrow==15.0.0
pandas==2.1.4
s3fs==2024.2.0
boto3==1.34.0
python-dotenv==1.0.0
structlog==23.2.0